# public/swagger.yaml
openapi: 3.0.0
info:
  description: API REST basado en Greenter - Facturación Electrónica.
  version: "1.3"
  title: Lycet API
  contact:
    name: Giancarlos Salas
    email: giansalex@gmail.com
...
security:
  - ApiToken: []
paths:
...
  /summary/send:
    post:
      tags:
        - summary
      summary: Enviar una Resumen diario de Boletas
      description: Retorna la respuesta de SUNAT y el XML firmado
      operationId: sendSummary
      requestBody:
        $ref: "#/components/requestBodies/Summary"
      responses:
        "200":
          description: Operacion exitosa
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SummaryResponse"
        "400":
          description: Error de Validación
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ValidationResponse"
  /summary/xml:
    post:
      tags:
        - summary
      summary: Generar el XML de un resumen diario
      description: Retorna el XML
      operationId: getSummaryXml
      requestBody:
        $ref: "#/components/requestBodies/Summary"
      responses:
        "200":
          description: Operacion exitosa
        "400":
          description: Error de Validación
          content:
            text/xml:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ValidationResponse"
  /summary/send_summary_xml:
    post:
      tags:
        - summary
      summary: Enviar el XML de un resumen diario generado
      description: Retorna la respuesta de SUNAT y el XML firmado
      operationId: sendSummaryXml
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FileDataSummaryXml"
        required: true
      responses:
        "200":
          description: Operacion exitosa
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DocumentResponse"
        "400":
          description: Error de Validación
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ValidationResponse"
  /summary/pdf:
    post:
      tags:
        - summary
      summary: Generar el PDF de un resumen diario
      description: Retorna el PDF
      operationId: getSummaryPdf
      requestBody:
        $ref: "#/components/requestBodies/Summary"
      responses:
        "200":
          description: Operacion exitosa
        "400":
          description: Error de Validación
          content:
            application/pdf:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ValidationResponse"
  /summary/status:
    get:
      tags:
        - summary
      summary: Obtiene el estado del ticket
      description: Retorna el XML
      operationId: getSummaryStatus
      parameters:
        - name: ticket
          in: query
          description: Ticket de Resumen diario
          schema:
            type: string
        - name: ruc
          in: query
          description: RUC de empresa (Multiples Empresas)
          required: false
          schema:
            type: string
      responses:
        "200":
          description: Operacion exitosa
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StatusResult"
  /voided/send:
    post:
      tags:
        - voided
      summary: Enviar una Comunicación de Bajas
      description: Retorna la respuesta de SUNAT y el XML firmado
      operationId: sendVoided
      requestBody:
        $ref: "#/components/requestBodies/Voided"
      responses:
        "200":
          description: Operacion exitosa
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SummaryResponse"
        "400":
          description: Error de Validación
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ValidationResponse"
  /voided/xml:
    post:
      tags:
        - voided
      summary: Generar el XML de una Comunicación de Bajas
      description: Retorna el XML
      operationId: getVoidedXml
      requestBody:
        $ref: "#/components/requestBodies/Voided"
      responses:
        "200":
          description: Operacion exitosa
        "400":
          description: Error de Validación
          content:
            text/xml:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ValidationResponse"
  /voided/pdf:
    post:
      tags:
        - voided
      summary: Generar el PDF de una Comunicación de Bajas
      description: Retorna el PDF
      operationId: getVoidedPdf
      requestBody:
        $ref: "#/components/requestBodies/Voided"
      responses:
        "200":
          description: Operacion exitosa
        "400":
          description: Error de Validación
          content:
            application/pdf:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ValidationResponse"
  /voided/status:
    get:
      tags:
        - voided
      summary: Obtiene el estado del ticket
      description: Retorna el XML
      operationId: getVoidedStatus
      parameters:
        - name: ticket
          in: query
          description: Ticket de Comunicación de baja
          schema:
            type: string
        - name: ruc
          in: query
          description: RUC de empresa (Multiples Empresas)
          required: false
          schema:
            type: string
      responses:
        "200":
          description: Operacion exitosa
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StatusResult"
...
servers:
  - url: http://lycet.api/api/v1
  - url: https://lycet.api/api/v1
components:
  requestBodies:
...
    Summary:
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Summary"
      description: Comprobante
      required: true
    Voided:
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Voided"
      description: Comprobante
      required: true
...
  securitySchemes:
    ApiToken:
      name: token
      type: apiKey
      in: query
      description: Api Token
  schemas:
    Address:
      type: object
      properties:
        ubigueo:
          type: string
        codigoPais:
          type: string
        departamento:
          type: string
        provincia:
          type: string
        distrito:
          type: string
        urbanizacion:
          type: string
        direccion:
          type: string
        codLocal:
          type: string
...
    Company:
      type: object
      properties:
        ruc:
          type: string
        razonSocial:
          type: string
        nombreComercial:
          type: string
        address:
          $ref: "#/components/schemas/Address"
        email:
          type: string
        telephone:
          type: string
...
    Document:
      type: object
      properties:
        tipoDoc:
          type: string
        nroDoc:
          type: string
...
    SummaryPerception:
      type: object
      properties:
        codReg:
          type: string
        tasa:
          type: number
          format: float
        mtoBase:
          type: number
          format: float
        mto:
          type: number
          format: float
        mtoTotal:
          type: number
          format: float
    SummaryDetail:
      type: object
      properties:
        tipoDoc:
          type: string
        serieNro:
          type: string
        clienteTipo:
          type: string
        clienteNro:
          type: string
        docReferencia:
          $ref: "#/components/schemas/Document"
        percepcion:
          $ref: "#/components/schemas/SummaryPerception"
        estado:
          type: string
        total:
          type: number
          format: float
        mtoOperGravadas:
          type: number
          format: float
        mtoOperInafectas:
          type: number
          format: float
        mtoOperExoneradas:
          type: number
          format: float
        mtoOperExportacion:
          type: number
          format: float
        mtoOperGratuitas:
          type: number
          format: float
        mtoOtrosCargos:
          type: number
          format: float
        mtoIGV:
          type: number
          format: float
        mtoIvap:
          type: number
          format: float
        mtoIcbper:
          type: number
          format: float
        mtoISC:
          type: number
          format: float
        mtoOtrosTributos:
          type: number
          format: float
    Summary:
      type: object
      properties:
        correlativo:
          type: string
        fecGeneracion:
          type: string
          format: date-time
        fecResumen:
          type: string
          format: date-time
        moneda:
          type: string
        company:
          $ref: "#/components/schemas/Company"
        details:
          type: array
          items:
            $ref: "#/components/schemas/SummaryDetail"
    VoidedDetail:
      type: object
      properties:
        tipoDoc:
          type: string
        serie:
          type: string
        correlativo:
          type: string
        desMotivoBaja:
          type: string
    Voided:
      type: object
      properties:
        correlativo:
          type: string
        fecGeneracion:
          type: string
          format: date-time
        fecComunicacion:
          type: string
          format: date-time
        company:
          $ref: "#/components/schemas/Company"
        details:
          type: array
          items:
            $ref: "#/components/schemas/VoidedDetail"
...
    Error:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
    CdrResponse:
      type: object
      properties:
        accepted:
          type: boolean
          readOnly: true
        id:
          type: string
        code:
          type: string
        description:
          type: string
        notes:
          type: array
          items:
            type: string
    BillResult:
      type: object
      properties:
        success:
          type: boolean
        error:
          $ref: "#/components/schemas/Error"
        cdrZip:
          type: string
          format: byte
        cdrResponse:
          $ref: "#/components/schemas/CdrResponse"
    SummaryResult:
      type: object
      properties:
        success:
          type: boolean
        error:
          $ref: "#/components/schemas/Error"
        ticket:
          type: string
    StatusResult:
      type: object
      properties:
        cdrZip:
          type: string
          format: byte
        cdrResponse:
          $ref: "#/components/schemas/CdrResponse"
        success:
          type: boolean
        error:
          $ref: "#/components/schemas/Error"
        code:
          type: string
    DocumentResponse:
      type: object
      properties:
        xml:
          type: string
        hash:
          type: string
          description: Resumen de la Firma Digital
        sunatResponse:
          $ref: "#/components/schemas/BillResult"
    SummaryResponse:
      type: object
      properties:
        xml:
          type: string
        hash:
          type: string
          description: Resumen de la Firma Digital
        sunatResponse:
          $ref: "#/components/schemas/SummaryResult"
    ValidationResponse:
      type: object
      properties:
        message:
          type: string
        field:
          type: string
...
    FileDataSummaryXml:
      type: object
      properties:
        filename:
          type: string
          description: Nombre de archivo
        ruc:
          type: string
          description: Ruc de la empresa